// src/lib/api/generated/model/eNotificationStatus.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * BASE API
 * OpenAPI spec version: 3.0
 */

/**
 * Giá trị: 0 = ACTIVE, 1 = INACTIVE
 */
export type ENotificationStatus = typeof ENotificationStatus[keyof typeof ENotificationStatus];


export const ENotificationStatus = {
  NUMBER_0: 0,
  NUMBER_1: 1,
} as const;