// src/lib/api/generated/model/eNotificationType.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * BASE API
 * OpenAPI spec version: 3.0
 */

/**
 * Giá trị: 0 = ALL, 1 = USER
 */
export type ENotificationType = typeof ENotificationType[keyof typeof ENotificationType];


export const ENotificationType = {
  NUMBER_0: 0,
  NUMBER_1: 1,
} as const;